# ml-matches: update and improve ambiguity computation