This uses the much-enhanced ambiguity computing algorithm in ml-matches, and implements this all better. The failures here are actually all type-intersection problems (https://github.com/JuliaLang/julia/issues/36951), as far as I can tell.