@timholy 2020-09-02
Sorry, I should have looked at this earlier. While you've carefully carved out exemptions for ambiguity tests that run within Julia itself, what should packages do? I have a lot of packages where the tests include `detect_ambiguities`, and as of today we're seeing new failures. For example, https://travis-ci.org/github/JuliaDebug/JuliaInterpreter.jl/jobs/723397378.

I'm not arguing that the new algorithm is wrong (I haven't checked), but just wondering whether there's a plan for moving forward.