@vtjnash 2020-09-02
There’s no reason anymore to include Base and Core in your list. List just the modules you care about.