@timholy 2020-09-02
Is that true even if we add methods to functions owned by Base?

I'll mimic the style of tests here. I am adding a "needs news" tag.