@tkf 2020-09-03
I noted above https://github.com/JuliaLang/julia/pull/36962#discussion_r483285985, I think we need more documentation. So, I'm also adding `needs docs`

I'm also adding `breaking` because:

1. It breaks existing user code bases.
2. If we retrofit a more specific API definition by introducing new documentation, it could be different from the users' mental models.  It is not ideal to change the behavior of a 4-year-old API (https://github.com/JuliaLang/julia/commit/be8679e965d89baf10161112e44f594339cd7daf) and claim a change in observable behavior as a bug fix or improvement when there is no documentation to ensure that the old and new behaviors are compatible in some ways.